requirements.txt
fastapi
uvicorn[standard]
pydantic

sqlalchemy
psycopg2-binary

python-dotenv

redis
rq

alembic

boto3

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

python-multipart
